requirements.txt
Django==4.2.2
django-apscheduler==0.6.0
feedparser==6.0.8
python-dateutil==2.8.2